requirements.txt
altair==5.1.1
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
filelock==3.12.4
gitdb==4.0.10
GitPython==3.1.37
htbuilder==0.6.2
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.26.0
opencv-python==4.8.1.78
openvino==2023.1.0
openvino-telemetry==2023.1.1
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
protobuf==4.24.3
pyarrow==13.0.0
pycodestyle==2.11.0
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
python-crfsuite==0.9.9
pytz==2023.3.post1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rich==13.6.0
rpds-py==0.10.3
six==1.16.0
smmap==5.0.1
st-annotated-text==4.0.1
streamlit==1.27.1
sympy==1.12
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
validators==0.22.0
watchdog==3.0.0
zipp==3.17.0
